md2ipynb-sync/requirements.txt
pyyaml>=6.0
jupytext>=1.16.0
nbformat>=5.9.0
watchdog>=3.0.0
PyQt5>=5.15.0
